DigitalCommunications

DigitalCommunications.ASK — Type

struct ASK <: AbstractScheme

Amplitude Shift Keying. The mapping rule is

\[ s_m = A_m g(t) cos(w_c t) \quad m = 1, \ldots, M\]

where $A_m = 2m - 1 - M$

DigitalCommunications.AWGNChannel — Type

mutable struct AWGNChannel

Additive white Gaussian noise channel.

Fields

snr::Float64

Channel SNR(Signal-to-Noise ratio)

DigitalCommunications.AbstractScheme — Type

abstract type AbstractScheme

Abstract type for modulation schemes such as PAM, ASK, PSK, FSK, QAM, etc.

DigitalCommunications.FSK — Type

struct FSK <: AbstractScheme

Frequency Shift Keying. The mapping rule is

\[ s_m = A cos(w_c t + w_m t) \quad m = 1, \ldots, M\]

where $w_m = m \Delta f$

DigitalCommunications.Generator — Type

struct Generator

Bit stream generator

Fields

bits::Array{Bool,1}

DigitalCommunications.GrayCoding — Type

struct GrayCoding <: AbstractCoding

Gray coding.

Note

In Gray coding just a single bit changes between adjacent symbols

Fields

pairs::Dict

DigitalCommunications.GrayCoding — Method

GrayCoding(k)

Constructs a GrayCoding with symbol size k.

Example

julia> gc = GrayCoding(2) 
GrayCoding(Dict([0, 1] => 2,[1, 1] => 3,[0, 0] => 1,[1, 0] => 4))

DigitalCommunications.Modulator — Type

struct Modulator{ST<:AbstractScheme, CT<:AbstractCoding}

Baseband digital modulator.

Fields

scheme::AbstractScheme

Modulation scheme

M::Int64

Constellation size

coding::AbstractCoding

Coding to map stream to code words

DigitalCommunications.PAM — Type

struct PAM <: AbstractScheme

Pulse Amplitude Modulation. The mapping rule is

\[ s_m = A_m p(t) \quad m = 1, \ldots, M\]

DigitalCommunications.PSK — Type

struct PSK <: AbstractScheme

Phase Shift Keying. The mapping rule is

\[ s_m = g(t) cos(w_c t - \theta_m) \quad m = 1, \ldots, M\]

where $\theta_m = \dfrac{2\pi(m - 1)}{M}$.

DigitalCommunications.QAM — Type

struct QAM <: AbstractScheme

Quadrature Amplitude Modulation. The mapping rule is

\[ s_m = A_{mi} g(t) cos(w_c t) - A_{mi} g(t) sin(w_c t) \]

where the amplitudes $A_{mi}, A_{mq} \in \{ \pm 1, \pm 3, \ldots, \pm (M - 1) \}$

DigitalCommunications.Q — Method

Q(x)

Q-function defined as.

\[ Q(x) = \int_{x}^{\infty} exp(-\dfrac{x^2}{2}) dx\]

DigitalCommunications.alphabet — Method

alphabet(modulator)

Returns the alphabet of the modulator modulator.

Example

julia> modulator = Modulator(PSK(), 4)  # 4-PSK modulator 
Modulator{PSK,GrayCoding}(PSK(), 4, GrayCoding(Dict([0, 1] => 2,[1, 1] => 3,[0, 0] => 1,[1, 0] => 4)))

julia> alphabet(modulator)
Dict{Array{Int64,1},Int64} with 4 entries:
  [0, 1] => 2
  [1, 1] => 3
  [0, 0] => 1
  [1, 0] => 4

DigitalCommunications.berask — Method

berask(γb, M)

Returns the probability of symbol error for the snr per bit γb and constallation size M for ASK and PAM signalling

DigitalCommunications.berfsk — Method

berfsk(γs, M)

Returns the probability of symbol error for the snr per bit γb and constallation size M for FSK signalling

DigitalCommunications.berpsk — Method

berpsk(γs, M)

Returns the probability of symbol error for the snr per bit γb and constallation size M for PSK signalling

DigitalCommunications.berqam — Method

berqam(γb, M)

Returns the probability of symbol error for the snr per bit γb and constallation size M for QAM signalling

DigitalCommunications.constellation — Method

constellation(modulator)

Plots the constellation diagram of the modulator.

DigitalCommunications.dbtosnr — Method

dbtosnr(γ)

Converts γ from dB value to its real value.

Example

julia> dbtosnr(2)  # γ = 2 dB
1.5848931924611136

DigitalCommunications.invmap — Method

invmap(coding)

Returns the inverse of the mapping. Inverse mapping maps the levels to bit chunks.

Example

julia> gray = GrayCoding(3)
GrayCoding(Dict([0, 1, 1] => 3,[0, 0, 1] => 2,[1, 1, 0] => 5,[0, 0, 0] => 1,[1, 0, 0] => 8,[1, 0, 1] => 7,[0, 1, 0] => 4,[1, 1, 1] => 6))

julia> invmap(gray)
Dict{Int64,Array{Int64,1}} with 8 entries:
  7 => [1, 0, 1]
  4 => [0, 1, 0]
  2 => [0, 0, 1]
  3 => [0, 1, 1]
  5 => [1, 1, 0]
  8 => [1, 0, 0]
  6 => [1, 1, 1]
  1 => [0, 0, 0]

DigitalCommunications.mapstream — Method

mapstream(modulator, stream)

Maps the bit stream into different levels according to the coding of the modulator modulator.

Example

julia> modulator = Modulator(PSK(), 4)  # 4-PSK modulator 
Modulator{PSK,GrayCoding}(PSK(), 4, GrayCoding(Dict([0, 1] => 2,[1, 1] => 3,[0, 0] => 1,[1, 0] => 4)))

julia> bits = rand(Bool, 10)
10-element Array{Bool,1}:
 0
 0
 0
 1
 1
 0
 1
 0
 0
 0

julia> mapstream(modulator, bits)
5-element Array{Int64,1}:
 1
 2
 4
 4
 1

DigitalCommunications.plotber — Method

plotber(; scheme, snr_per_bit_range, krange, pltkwargs...)

Plots the probability of symbols error versus snr per bit for the signaling scheme. snr_per_bit_range is the snr per bit range and krange is the symbol size of the M-ary signalling where $M=2^k$.

DigitalCommunications.scheme — Method

Returns scheme of the modulation modulator.

DigitalCommunications.signalset — Method

signalset(modulator)

Returns the signal set of the modulator

Example

julia> modulator = Modulator(PSK(), 4)
Modulator{PSK,GrayCoding}(PSK(), 4, GrayCoding(Dict([0, 1] => 2,[1, 1] => 3,[0, 0] => 1,[1, 0] => 4)))

julia> signalset(modulator)
4-element Array{Array{Float64,1},1}:
 [1.0, 0.0]
 [6.123233995736766e-17, 1.0]
 [-1.0, 1.2246467991473532e-16]
 [-1.8369701987210297e-16, -1.0]

DigitalCommunications.symbolsize — Method

symbolsize(modulator)

Returns the symbols size of the modulator modulator.